Parallel loops: Partitioning Added the 'Partitioning' class to the Parallel Loops project.
This commit introduces a new example to the 'Parallel Loops' project, a 'Partitioning' class that demonstrates the use of the 'Partitioner' class for efficient handling of parallel computations. It was necessary to add this example to convey the benefits of partitioning large data sets for parallel processes. Also, added the BenchmarkDotNet package for performance testing of these methods.