Parallel LINQ: AsParallel and ParallelQuery
Created a new ParallelProgramming.ParallelLINQ project to explore the usage of AsParallel and ParallelQuery features in Parallel LINQ. The main Program.cs contains an instance of AsParallelAndParallelQuery that demonstrates the transformation of an IEnumerable into a ParallelQuery which enables parallelised operations. This new project has been added to the NET Examples solution. An example, AsParallelAndParallelQuery, is created to experiment with different operations on parallelised collections.