Additional Lectures: Duck Typing Mixins Added a new example demonstrating Duck Typing and Mixins.
The example was part of the additional lecture series and offers a live coding demo to help understand these concepts better. The files were included in the DesignPatterns.Additional project, so it was necessary to update the solution file as well. The solution file has been updated to include the new project.